feat: write audit results to disk on failure (#68)
Pulled from #32, this should make it easier to understand why a failure
happened and can be especially useful when there's inconsistent results.

I was hoping to have something similar to Capybara failure screenshots
in Rails apps (which end up in/named e.g.
`failures_r_spec_example_groups_homepage_passes_a_lighthouse_accessibility_audit_808`)
but it turns out that is a Rails feature that seems to involve capturing
the example name (or `method_name`, as its called by their code)
separately when the suite starts, and while we look to have access to
that in `matchers.rb` it just felt best to start simple.

That is also why for now I've not had the results directory use
`Rails.root.join(...)` by default, though that will probably be the next
thing I add

Closes #32