feat: Add support for passing "preset" and "form-factor" values to lighthouse (#71)
This PR adds config options for preset and form factor so these values
can be passed to lighthouse. This allows these matchers to be used for
apps that are desktop or mobile only.

More information about these options can be found on the main lighthouse
repo: https://github.com/GoogleChrome/lighthouse